Move consolidate tests to testutil
